format added to Property object, split class.
