+++ Directory.Build.props
@@ -1,5 +1,5 @@
 <Project>
   <PropertyGroup>
-    <Version>0.3.10</Version>
+    <Version>0.3.11</Version>
   </PropertyGroup>
 </Project>
